Add cell hyperlinks to other params with history
Grids now support cells which are clickable links.

Selecting any param (left grid) with the mouse counts as clicking its
link.

Yapped now maintains a history of link clicks. The user can go back and
forward through the history with the usual mouse buttons, or the new
back/forward toolbar buttons.

History is persistent across sessions.

Added additional layout information for params, unique to Yapped, in
JSON format.

This includes an optional alternate display name for a cell. For
example, Magic's ezStateBehaviorType is the type of spell tool used to
cast the spell, and so could be displayed as "ezStateBehaviorType
[spellTool]".

This also includes set of conditions under which a cell value may become
a link to a row in another param.

For example, in Dark Souls 3, Magic's refId1 value is a reference to
either:

a) a BehaviorParam_PC if refCategory1 is 0;
b) a Bullet if refCategory1 is 1; or
c) an SpEffectParam if refCategory1 is 2.

Conditions are optional. For example, in Dark Souls 3,
EquipParamWeapon's correctType value is always a reference to a
CalcCorrectGraph.