Improve the process of adding and deleting rows
If the user enters an ID which already exists, allow them to edit the
value they just set rather than their having to ask to duplicate/add a
row again (and having to remember what they typed).

When duplicating a row, start by prompting the user with the ID and
name of the row they're duplicating. Row IDs may be large and long and
counting zeroes to make adjacent IDs is tiresome.

Use Button DialogResults rather than reimplementing their functionality.

Apply a fix to ensure the new grids update when a row is deleted.

Avoid using MessageBoxIcon.Question which is deprecated.