Add cell details panel and graph CalcCorrectGraph
Add a cell details panel below the cells grid, currently only showing
its name and the description which already appears on the cell tooltip.

When viewing CalcCorrectGraph params, this details view switches to a
simple graph which plots the curve described by the param row. Since
it's far easier to parse a visual graph than a set of numbers, this
helps when editing.

Update our CalcCorrectGraph calculator, as used in the weapon damage
calculator tool and now this new graph view, with hopefully more correct
calculations. Largely based on intuition since the plotted graphs don't
make as much sense using the old method.